refactor getRememberMeHash(req) so it can be used by hub and next
